🐛 Fix weird event history bug.
So for some reason pghistory didn't correctly see the old migration as needing the trigger remove/readd. But if I split it into two migrations (column remove and column add) then it works. This technically means data loss but it will rapidly rescrape and I wasn't use that data for anything anyway.